requirements.txt
pandas==1.4.3
numpy==1.19.2
matplotlib==3.3.2
fpdf2==2.5.7
PyPDF2==2.11.1